requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.4=py310h8d17308_0
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
blosc=1.21.3=hdccc3a2_0
boost-cpp=1.78.0=h9f4b32c_1
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
brotlipy=0.7.0=py310h8d17308_1005
bzip2=1.0.8=he774522_0
c-ares=1.18.1=h8ffe710_0
ca-certificates=2023.5.7=h56e8100_0
cachetools=5.3.0=pyhd8ed1ab_0
cairo=1.16.0=hd694305_1014
ceres-solver=2.1.0=hb02fd38_1
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py310h2bbff1b_3
cfitsio=4.2.0=h9ebe7e4_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py310h232114e_0
coverage=7.2.1=py310h8d17308_0
cryptography=39.0.2=py310h6e82f81_0
curl=7.88.1=h68f0423_0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
draco=1.5.6=h181d51b_0
eigen=3.4.0=h2d74725_0
exiv2=0.27.6=hadc2d18_1
expat=2.5.0=h1537add_0
flatbuffers=23.3.3=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.39.0=py310h8d17308_0
freetype=2.12.1=ha860e81_0
freexl=1.0.6=h67ca5e6_1
frozenlist=1.3.3=py310h8d17308_0
future=0.18.3=py310haa95532_0
gast=0.4.0=pypi_0
gdal=3.6.2=py310h644bc08_8
geos=3.11.1=h1537add_0
geotiff=1.7.1=hc256dc0_6
gettext=0.21.1=h5728263_0
gflags=2.2.2=ha925a31_1004
giflib=5.2.1=h8cc25b3_3
glib=2.74.1=h12be248_1
glib-tools=2.74.1=h12be248_1
glog=0.6.0=h4797de2_0
google-auth=2.17.3=pyh1a96a4e_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.0=py310h76bb054_0
gsl=2.7=hdfb1a43_0
gst-plugins-base=1.22.0=h001b923_2
gstreamer=1.22.0=h6b5321d_2
h5py=3.8.0=pypi_0
hdf4=4.2.15=h1b1b6ef_5
hdf5=1.12.2=nompi_h57737ce_101
httplib2=0.21.0=pyhd8ed1ab_0
icc_rt=2022.1.0=h6049295_2
icu=70.1=h0e60522_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.6.0=pyha770c72_0
intel-openmp=2021.4.0=haa95532_3556
jax=0.4.8=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.1=py310haa95532_0
jpeg=9e=h2bbff1b_1
kealib=1.5.0=h61be68b_0
keras=2.12.0=pypi_0
kiwisolver=1.4.4=py310h232114e_1
krb5=1.20.1=heb0366b_0
laz-perf=3.3.0=h91493d7_0
lcms2=2.15=ha5c8aab_0
lerc=4.0.0=h63175ca_0
libabseil=20230125.0=cxx17_h63175ca_1
libaec=1.0.6=h63175ca_1
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=12_win64_mkl
libclang=16.0.0=pypi_0
libclang13=15.0.7=default_h77d9078_1
libcurl=7.88.1=h68f0423_0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libgdal=3.6.2=h221c0cb_8
libglib=2.74.1=he8f3873_1
libgrpc=1.54.0=h32da247_0
libhwloc=2.9.0=h51c2c0f_0
libiconv=1.17=h8ffe710_0
libkml=1.3.0=hf2ab4e4_1015
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libnetcdf=4.9.1=nompi_hc41bf00_101
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libpq=15.2=ha9684e8_0
libprotobuf=3.21.12=h12be248_0
librttopo=1.1.0=he22b5cd_12
libspatialindex=1.9.3=h39d44d4_4
libspatialite=5.0.1=hfdcade0_23
libsqlite=3.40.0=hcfcfb64_0
libssh2=1.10.0=h9a1e1f7_3
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=hc3477c8_0
libxslt=1.1.37=h0192164_0
libzip=1.9.2=h519de47_1
libzlib=1.2.13=hcfcfb64_4
lxml=4.9.2=py310hc0e5b84_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.3=pyhd8ed1ab_0
markupsafe=2.1.2=py310h8d17308_0
matplotlib=3.7.1=py310h5588dad_0
matplotlib-base=3.7.1=py310h51140c5_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
ml-dtypes=0.1.0=pypi_0
mock=5.0.1=pyhd8ed1ab_0
multidict=6.0.4=py310h8d17308_0
munkres=1.1.4=pyh9f0ad1d_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
nitro=2.7.dev6=h1537add_5
nose2=0.9.2=py_0
numpy=1.23.5=py310h60c9a35_0
numpy-base=1.23.5=py310h04254f7_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=ha2aaf27_2
openssl=3.1.0=hcfcfb64_2
opt-einsum=3.3.0=pypi_0
owslib=0.28.1=pyhd8ed1ab_0
packaging=23.0=pyhd8ed1ab_0
pandas=2.0.1=pypi_0
pcre2=10.40=h17e33f8_0
pdal=2.5.2=h2d23c47_1
pillow=9.4.0=py310hd77b12b_0
pip=23.0.1=py310haa95532_0
pixman=0.40.0=h8ffe710_0
plotly=5.13.1=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.4.0=pyhd3eb1b0_0
poppler=23.03.0=h183ae7b_0
poppler-data=0.4.12=hd8ed1ab_0
postgresql=15.2=hd87cd2b_0
proj=9.1.1=heca977f_2
protobuf=4.21.12=py310h00ffb61_0
psycopg2=2.9.3=py310h709a2f1_2
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyproj=3.4.1=py310hcdd211a_1
pyqt=5.15.7=py310h1fd54f2_3
pyqt5-sip=12.11.0=py310h00ffb61_3
pyqtwebkit=5.15.9=py310he49db7d_1
pysocks=1.7.1=pyh0701188_6
python=3.10.9=h4de0772_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=1.12.1=cpu_py310h5e1f01c_1
pytz=2022.7.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py310h2bbff1b_1
qca=2.3.5=h2624d1c_3
qgis=3.30.0=py310hc91cd07_0
qjson=0.9.0=h04a78d6_1009
qscintilla2=2.13.4=py310he49db7d_0
qt-main=5.15.8=h720456b_6
qtkeychain=0.13.2=h04a78d6_2
qtwebkit=5.212=h992fabe_8
qwt=6.2.0=h07be427_5
re2=2023.02.02=h63175ca_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.2.2=py310hd77b12b_0
scipy=1.10.1=py310hb9afe5d_0
setuptools=65.6.3=py310haa95532_0
sip=6.7.7=py310h00ffb61_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=hfb803bf_0
sqlite=3.40.1=h2bbff1b_0
suitesparse=5.4.0=h5d0cbe0_1
tbb=2021.8.0=h91493d7_0
tenacity=8.2.2=pyhd8ed1ab_0
tensorboard=2.12.2=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py310h5588dad_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tikzplotlib=0.10.1=pyh050c7b8_0
tiledb=2.13.2=h3132609_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torch-tb-profiler=0.4.1=pypi_0
torchvision=0.13.1=cpu_py310h378ed51_0
tornado=6.2=py310h8d17308_1
typing-extensions=4.4.0=py310haa95532_0
typing_extensions=4.4.0=py310haa95532_0
tzdata=2023.3=pypi_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.0.0=py310h8d17308_0
urllib3=1.26.14=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.34.31931=h4c5c07a_10
webcolors=1.13=pyhd8ed1ab_0
werkzeug=2.3.0=pyhd8ed1ab_0
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wincertstore=0.2=py310haa95532_2
wrapt=1.14.1=pypi_0
xerces-c=3.2.4=h63175ca_1
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=he774522_0
yarl=1.9.1=py310h8d17308_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h19a0ad4_0
